create User Repository - use JpaSpecification to filter query on hasDevice
Using JpaSpecification instead of query methods in the interface as we anticipate more filters on the User repository in the future - and do not want to have to define many different repository methods with all the different filters.

Create test samples for all the entities.

Refactor package structure to nest all data related classes in {domain}.data instead of directly in the domain package.